add the uplode file for post
