+++ test/youpi.conf
@@ -8,7 +8,7 @@ server {
 	client_max_body_size 10M
 
 
-	allow_methods GET DELETE
+	allow_methods GET DELETE POST
 	error_page 403 /404.py
 
 	location /404.html
